# Backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/NMK": {
            "get": {
                "description": "Fetch all NMK codes, with optional filtering based on query parameters such as email, role, and NMK code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NMK"
                ],
                "summary": "Retrieve all NMK codes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by user email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by role; requires superadmin privileges to access unverified NMK codes",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by NMK Code",
                        "name": "NMK_Code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved NMK codes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.NMK"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid NMK code",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to fetch NMK codes",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/NMK/approve": {
            "post": {
                "description": "Approves an NMK code, setting its IsVerified status to true.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NMK"
                ],
                "summary": "Approve an existing NMK code by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the NMK code to approve",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully approved NMK code",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid NMK code ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to approve NMK code",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/addDoctor": {
            "post": {
                "description": "Add a new doctor to the system with the details provided in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Add a new doctor",
                "parameters": [
                    {
                        "description": "Doctor data",
                        "name": "doctor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DoctorData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully added doctor",
                        "schema": {
                            "$ref": "#/definitions/models.DoctorData"
                        }
                    },
                    "400": {
                        "description": "Failed to parse request body",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to insert doctor",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
        "/addNMK": {
            "post": {
                "description": "Creates a new NMK code in the repository. The new NMK code is initially marked as unverified.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NMK"
                ],
                "summary": "Add a new NMK code",
                "parameters": [
                    {
                        "description": "New NMK Code details",
                        "name": "nmk",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NMK"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully added NMK code",
                        "schema": {
                            "$ref": "#/definitions/models.NMK"
                        }
                    },
                    "400": {
                        "description": "Failed to bind NMK data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to insert NMK data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/doctor": {
            "get": {
                "description": "Retrieve all doctors, with filtering options based on query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Retrieve all doctors",
                "parameters": [
                    {
                        "enum": [
                            "superadmin",
                            "admin",
                            "user"
                        ],
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Doctor Code",
                        "name": "Doctor_Code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DoctorData"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch doctors",
                        "schema": {
                            "$ref": "#/definitions/responses.ApplicationResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.DoctorData": {
            "type": "object",
            "properties": {
                "ClinicAddress": {
                    "type": "string"
                },
                "ContactNumber": {
                    "type": "integer"
                },
                "Description": {
                    "type": "string"
                },
                "Docter_Code": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "ImageURL": {
                    "type": "string"
                },
                "IsVerified": {
                    "type": "boolean"
                },
                "Name": {
                    "type": "string"
                },
                "Specialization": {
                    "type": "string"
                },
                "_id": {
                    "type": "string"
                }
            }
        },
...
        "models.NMK": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "Contact_Number": {
                    "type": "integer"
                },
                "District": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "Established_Year": {
                    "type": "integer"
                },
                "ImageURL": {
                    "type": "string"
                },
                "IsVerified": {
                    "type": "boolean"
                },
                "NMK_Code": {
                    "type": "string"
                },
                "NMK_Image": {
                    "type": "string"
                },
                "NMK_Verification_Image": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Owner_Name": {
                    "type": "string"
                },
                "Pincode": {
                    "type": "integer"
                },
                "State": {
                    "type": "string"
                },
                "_id": {
                    "type": "string"
                }
            }
        },
...
        "responses.ApplicationResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {}
            }
        }
    }
}